% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbLab.R
\name{herbLab}
\alias{herbLab}
\title{Make herbarium labels}
\usage{
herbLab(fieldbook = NULL,
        dir_create = "results_herbarium_labels",
        file_label = "herbarium_labels.pdf")
}
\arguments{
\item{fieldbook}{The CSV formatted file containing all details about the collected
specimens. See the README file for a complete description on how the columns of
the input field book should be formatted.}

\item{dir_create}{Path to the directory where the labels(s) will be saved. The
default setting creates a directory named \strong{results_herbarium_labels} where
the labels will be saved in a subfolder named by the current date.}

\item{file_label}{Name of the resulting PDF formatted herbarium label files.
If no name is reported, the default setting creates files named as
\emph{herbarium_labels_1.pdf}, \emph{herbarium_labels_2.pdf}, etc, depending on how many
labels you need to prepare, because the function put six labels in each PDF file.}
}
\description{
This function makes herbarium labels from field book in spreadsheet
CSV format. It is currently more comprehensive for making herbarium labels from
specimens collected in the USA, because it will display geographic maps also
at county level. For specimens collected in all other countries, the label will
display only the country level map. If geographic coordinates are provided,
then the specimen record is also plotted in the map. The function also insert
taxon authorities and nomenclatural updates automatically, by implementing an
internal function from \href{https://idiv-biodiversity.github.io/lcvplants/}{lcvplants}
package.
}
\author{
Domingos Cardoso
}
